Backporting hibernate upgrade to 1.9.x branch - SEARCH-10
git-svn-id: http://svn.openmrs.org/openmrs/branches/1.9.x@25042 5bac5841-c719-aa4e-b3fe-cce5062f897a

+++ api/src/main/java/org/openmrs/api/db/hibernate/HibernateEncounterDAO.java
@@ -324,11 +324,11 @@ public class HibernateEncounterDAO implements EncounterDAO {
 	 * @see org.openmrs.api.db.EncounterDAO#getCountOfEncounters(java.lang.String, boolean)
 	 */
 	@Override
-	public Integer getCountOfEncounters(String query, boolean includeVoided) {
+	public Long getCountOfEncounters(String query, boolean includeVoided) {
 		Criteria criteria = createEncounterByQueryCriteria(query, includeVoided, false);
 		
 		criteria.setProjection(Projections.countDistinct("enc.encounterId"));
-		return (Integer) criteria.uniqueResult();
+		return (Long) criteria.uniqueResult();
 	}
 	
 	/**
